+++ feed-consumer/Ingester.fs
@@ -1,5 +1,6 @@
 module FeedConsumerTemplate.Ingester
 
+open Propulsion.Internal
 open System
 open FeedConsumerTemplate.Domain
 
@@ -39,7 +40,7 @@ module PipelineEvent =
             ValueSome (fc, s |> Seq.map (fun e -> Unchecked.unbox<Item> e.Context))
         | _ -> ValueNone
 
-let handle maxDop struct (stream, span) = async {
+let handle maxDop stream span ct = Async.startImmediateAsTask ct <| async {
     match stream, span with
     | PipelineEvent.ItemsForFc (fc, items) ->
         // Take chunks of max 1000 in order to make handler latency be less 'lumpy'