+++ propulsion-projector/Handler.fs
@@ -74,7 +74,7 @@ let categoryFilter = function
     | "categoryA"
     | _ -> true
 
-let handle struct (_stream, span: Propulsion.Streams.StreamSpan<_>) = async {
+let handle _stream (span: Propulsion.Streams.StreamSpan<_>) _ct = task {
     let r = System.Random()
     let ms = r.Next(1, span.Length)
     do! Async.Sleep ms
@@ -84,8 +84,8 @@ let handle struct (_stream, span: Propulsion.Streams.StreamSpan<_>) = async {
 type Config private () =
     
     static member StartSink(log : Serilog.ILogger, stats,
-                            handle : struct (FsCodec.StreamName * Propulsion.Streams.Default.StreamSpan)
-                                     -> Async<struct (Propulsion.Streams.SpanResult * 'Outcome)>,
+                            handle : System.Func<FsCodec.StreamName, Propulsion.Streams.Default.StreamSpan, _,
+                                     System.Threading.Tasks.Task<struct (Propulsion.Streams.SpanResult * 'Outcome)>>,
                             maxReadAhead : int, maxConcurrentStreams : int, ?wakeForResults, ?idleDelay, ?purgeInterval) =
         Propulsion.Streams.Default.Config.Start(log, maxReadAhead, maxConcurrentStreams, handle, stats, stats.StatsInterval.Period,
                                                 ?wakeForResults = wakeForResults, ?idleDelay = idleDelay, ?purgeInterval = purgeInterval)